Fixed segfault in test while running CTest
Running tests directly from the executable would result in all tests
passing and a clean exit, but CTest would fail to execute all of "Can
instantiate compounds" and would segfault. Reason was that `multiType`
was being used in `nestedType` but only `nestedType` was registered in
that test case. When running all tests in one go this was not an
error (a previous test had registered it) but as CTest runs each test
sequentially in a different process, the `multiType` had not been
registered.